The perf_events code was using __u64 instead of uint64_t and this was causing a warning when compiling for 64-bit Power.

+++ src/papi_pfm_events.c
@@ -1792,7 +1792,7 @@ long long generate_p4_event(long long escr,
 }
 
 int
-_papi_pfm3_setup_counters( unsigned long long *pe_event, hwd_register_t *ni_bits ) {
+_papi_pfm3_setup_counters( uint64_t *pe_event, hwd_register_t *ni_bits ) {
 
   int ret;
